ExchangeRate record created

+++ src/main/java/software/ulpgc/moneycalculator/Money.java
@@ -1,4 +1,4 @@
 package software.ulpgc.moneycalculator;
 
-public record Money(long amount, Currency currency) {
+public record Money(double amount, Currency currency) {
 }
